django-rest-framework-api/requirements.txt
djangorestframework==3.6.2
Django==2.1.7
Pygments==2.1.3
Markdown==2.6.3
musicbrainzngs==0.5
